+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 # Unreleased
 
+- Added Flash support via Ruffle (https://github.com/funmaker/Hybooru/issues/34)
 - Added blurhash support
 - Added Hydrus version check
 - Fixed thumbnail fading